back/pyproject.toml
[tool.poetry]
name = "back"
version = "0.1.0"
description = ""
authors = ["MrWhiteres <alexandr.leluk.99@gmail.com>"]
readme = "../README.md"

[tool.poetry.dependencies]
python = "3.11.7"
django = "5.0.1"
daphne = "4.0.0"
uvicorn = "0.27.0"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.3.1"
django-cors-headers = "4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
